// wrapper/Enums/DOTAScriptInventorySlot_t.ts
export enum DOTAScriptInventorySlot_t {
	DOTA_ITEM_SLOT_1 = 0,
	DOTA_ITEM_SLOT_2 = 1,
	DOTA_ITEM_SLOT_3 = 2,
	DOTA_ITEM_SLOT_4 = 3,
	DOTA_ITEM_SLOT_5 = 4,
	DOTA_ITEM_SLOT_6 = 5,
	DOTA_ITEM_SLOT_7 = 6,
	DOTA_ITEM_SLOT_8 = 7,
	DOTA_ITEM_SLOT_9 = 8,
	DOTA_STASH_SLOT_1 = 9,
	DOTA_STASH_SLOT_2 = 10,
	DOTA_STASH_SLOT_3 = 11,
	DOTA_STASH_SLOT_4 = 12,
	DOTA_STASH_SLOT_5 = 13,
	DOTA_STASH_SLOT_6 = 14,
	DOTA_ITEM_TP_SCROLL = 15,
	DOTA_ITEM_NEUTRAL_SLOT = 16,
	DOTA_ITEM_TRANSIENT_ITEM = 17,
	DOTA_ITEM_TRANSIENT_RECIPE = 18,
	DOTA_ITEM_TRANSIENT_CAST_ITEM = 20,
}